Create spark ui adapter
