+++ piezo_web_app/PiezoWebApp/src/services/spark_job/i_spark_ui_adapter.py
@@ -1,7 +1,7 @@
 from abc import ABCMeta, abstractmethod
 
 
-class ISparkUiService(metaclass=ABCMeta):
+class ISparkUiAdapter(metaclass=ABCMeta):
     @staticmethod
     def create_ui_proxy_body(job_name, namespace):
         pass
